refactor(image-uploader): separate shared logic from server-only logic
