+++ src/features/image-uploader/apis/index.ts
@@ -1,7 +1,1 @@
-import * as UploadImageApi from './upload-image.api'
-import * as UploadImageRoute from './upload-image.route'
-
-export const UploadImage = {
-  route: UploadImageRoute,
-  api: UploadImageApi,
-}
+export * from './upload-image'
