+++ src/features/image-uploader/route/index.ts
@@ -1,0 +1,1 @@
+export * as UploadImage from './upload-image.route'
