+++ src/features/image-uploader/route/upload-image.route.ts
