+++ src/features/image-uploader/server/index.ts
@@ -1,0 +1,2 @@
+export * from '..'
+export * from '../route'
